# packages/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Town Backend",
    "description": "二次元 AI 小镇陪伴智能体 - World Engine + LLM",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "List Notifications",
        "description": "获取通知列表\n\nArgs:\n    limit: 返回数量（最大 200）\n    unread_only: 仅返回未读通知\n\nReturns:\n    通知列表（按时间倒序，最新的在前）",
        "operationId": "list_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "unread_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Unread Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "notifications"
        ],
        "summary": "Create Notification",
        "description": "手动创建通知（前端\"模拟通知\"按钮调用）\n\nBody:\n    type: 通知类型 (share/system/character/qq)\n    title: 标题\n    content: 内容",
        "operationId": "create_notification_api_v1_notifications_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationCreatedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Clear All Notifications",
        "description": "清除所有通知",
        "operationId": "clear_all_notifications_api_v1_notifications_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsClearedOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notif_id}/read": {
      "put": {
        "tags": [
          "notifications"
        ],
        "summary": "Mark Notification Read",
        "description": "标记单条通知为已读",
        "operationId": "mark_notification_read_api_v1_notifications__notif_id__read_put",
        "parameters": [
          {
            "name": "notif_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notif Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationMarkedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/{notif_id}": {
      "delete": {
        "tags": [
          "notifications"
        ],
        "summary": "Delete Notification",
        "description": "删除单条通知",
        "operationId": "delete_notification_api_v1_notifications__notif_id__delete",
        "parameters": [
          {
            "name": "notif_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notif Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDeletedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Login",
        "description": "登录接口 - 账号密码换取 JWT Token\n\n请求体: {\"username\": \"admin\", \"password\": \"admin123\"}\n返回: {\"token\": \"jwt_token\", \"user_id\": \"admin\", \"expires_in\": 86400}",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/duration/calculate": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Calculate Duration",
        "description": "计算动态耗时\n\nArgs:\n    base_duration: 基础耗时（分钟）\n    weather: 天气\n    is_outdoor: 是否户外\n    crowdedness: 拥挤度 0-1\n    stamina: 体力 0-100\n    mood: 情绪",
        "operationId": "calculate_duration_api_v1_duration_calculate_get",
        "parameters": [
          {
            "name": "base_duration",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Base Duration"
            }
          },
          {
            "name": "weather",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "sunny",
              "title": "Weather"
            }
          },
          {
            "name": "is_outdoor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Is Outdoor"
            }
          },
          {
            "name": "crowdedness",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.0,
              "title": "Crowdedness"
            }
          },
          {
            "name": "stamina",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Stamina"
            }
          },
          {
            "name": "mood",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "calm",
              "title": "Mood"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DurationCalculateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppNotificationOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "read": {
            "type": "boolean",
            "title": "Read"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "title",
          "content",
          "created_at",
          "read"
        ],
        "title": "AppNotificationOut"
      },
...
      "DurationCalculateOut": {
        "properties": {},
        "additionalProperties": true,
        "type": "object",
        "title": "DurationCalculateOut",
        "description": "动态耗时计算结果（结构随移动矩阵输出，暂透传）"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginOut": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "token",
          "user_id"
        ],
        "title": "LoginOut"
      },
...
      "NotificationCreatedOut": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AppNotificationOut"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "NotificationCreatedOut"
      },
      "NotificationDeletedOut": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "id"
        ],
        "title": "NotificationDeletedOut"
      },
      "NotificationMarkedOut": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "id"
        ],
        "title": "NotificationMarkedOut"
      },
      "NotificationsClearedOut": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "NotificationsClearedOut"
      },
      "NotificationsListOut": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AppNotificationOut"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "unread": {
            "type": "integer",
            "title": "Unread"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "unread"
        ],
        "title": "NotificationsListOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}